Just have scripts folder as functions
