+++ nbsite/scripts/__init__.py
@@ -1,0 +1,2 @@
+from ._fix_links import fix_links
+from ._clean_dist_html import clean_dist_html
