--- scripts/nbsite_from_tmplate.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-
-# TODO: deprecate; tell people to use nbsite init
-import sys
-
-from nbsite.cmd import init
-
-init(sys.argv[1])
